requirement.txt
apache-airflow
duckdb
pandas
kubernetes